+++ scripts/balloon-trip-obstacle.gd
@@ -1,0 +1,3 @@
+extends ReferenceRect
+var path:String
+var number:int = 0